Refactor human results handling and streamline captain benchmark execution
- Removed the get_human_results function from run_captain_benchmarks.py to simplify the codebase and reduce dependencies on pandas.
- Updated run_captain_benchmarks.py to enhance logging and improve the display of progress during experiments.
- Adjusted argument parsing to remove the --include-human flag, reflecting the removal of human results processing.
- Added the get_human_results function to analysis.py for better organization and potential future use.
